+++ src/app.c
@@ -28,8 +28,9 @@ App *initialize_app(int argc, char **argv)
     // Connect Greeter & UI Signals
     g_signal_connect(app->greeter, "authentication-complete",
                      G_CALLBACK(authentication_complete_cb), app);
-    g_signal_connect(GTK_ENTRY(APP_PASSWORD_INPUT(app)), "activate",
-                     G_CALLBACK(handle_password), app);
+    app->password_callback_id =
+        g_signal_connect(GTK_ENTRY(APP_PASSWORD_INPUT(app)), "activate",
+                         G_CALLBACK(handle_password), app);
     // This was added to fix a bug where the background window would be focused
     // instead of the main window, preventing users from entering their password.
     // It's undocument & probably not necessary any more. Investigate & remove.